+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Shopify
 
+## Unreleased
+
+- Added `craft\shopify\elements\db\ProductQuery::publishedScope()`. ([#65](https://github.com/craftcms/shopify/issues/65))
+
 ## 3.1.1 - 2023-01-20
 
 - Fixed a SQL error that occurred when syncing products with several tags. ([#54](https://github.com/craftcms/shopify/issues/54))